Move Variables-orientation utils into Domain dir
With this change, the DataStructures no longer depend on the Domain.
