+++ tests/Unit/DataStructures/CMakeLists.txt
@@ -18,7 +18,6 @@ set(LIBRARY_SOURCES
   Test_IndexIterator.cpp
   Test_LeviCivitaIterator.cpp
   Test_ModalVector.cpp
-  Test_OrientVariables.cpp
   Test_SliceIterator.cpp
   Test_SpinWeighted.cpp
   Test_StripeIterator.cpp
